Refactor to fit in with Vercel
